Suppress warnings per gcc 5.4.0

+++ include/aikido/statespace/detail/Rn-impl.hpp
@@ -146,7 +146,7 @@ template <int N>
 void R<N>::setValue(State* _state, const typename R<N>::Vectord& _value) const
 {
   // TODO: Skip this check in release mode.
-  if (_value.size() != getDimension())
+  if (static_cast<std::size_t>(_value.size()) != getDimension())
   {
     std::stringstream msg;
     msg << "Value has incorrect size: expected " << getDimension() << ", got "
@@ -248,7 +248,7 @@ void R<N>::expMap(
     const Eigen::VectorXd& _tangent, StateSpace::State* _out) const
 {
   // TODO: Skip this check in release mode.
-  if (_tangent.size() != getDimension())
+  if (static_cast<std::size_t>(_tangent.size()) != getDimension())
   {
     std::stringstream msg;
     msg << "Tangent vector has incorrect size: expected " << getDimension()
@@ -264,7 +264,7 @@ void R<N>::expMap(
 template <int N>
 void R<N>::logMap(const StateSpace::State* _in, Eigen::VectorXd& _tangent) const
 {
-  if (_tangent.size() != getDimension())
+  if (static_cast<std::size_t>(_tangent.size()) != getDimension())
     _tangent.resize(getDimension());
 
   auto in = static_cast<const State*>(_in);
